pyproject.toml
[tool.poetry]
name = "n5geh.tools.adapters.glendimplex"
version = "0.1.0"
description = ""
authors = ["StWiemann <stephan.wiemann@tu-dresden.de>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
paho-mqtt = "^1.5.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
